+++ lib/active_record/connection_adapters/sqlserver/type/decimal.rb
@@ -4,7 +4,6 @@ module ActiveRecord
       module Type
         class Decimal < ActiveRecord::Type::Decimal
 
-          include Castable
 
         end
       end