+++ lib/active_record/connection_adapters/sqlserver/type/float.rb
@@ -4,8 +4,6 @@ module ActiveRecord
       module Type
         class Float < ActiveRecord::Type::Float
 
-          include Castable
-
           def type
             :float
           end