+++ lib/active_record/connection_adapters/sqlserver/type/real.rb
@@ -4,8 +4,6 @@ module ActiveRecord
       module Type
         class Real < Float
 
-          include Castable
-
           def type
             :real
           end