+++ lib/active_record/connection_adapters/sqlserver/type/smalldatetime.rb
@@ -4,8 +4,6 @@ module ActiveRecord
       module Type
         class SmallDateTime < DateTime
 
-          include Castable
-
 
           private